Rename HTMLUtils.sanitize to HTMLUtils.cleanse
"sanitize" more often refers to filtering content for security
purposes, which we're now doing in new versions of HTMLUtils whereas
"cleanse", henceforth, will refer to the general process of munging,
formatting, sanitizing, et al.
